feat: mixed expressions, initial implementation. Complete rehaul (WIP)
