+++ src/api/enums.ts
@@ -1,0 +1,5 @@
+export enum EventType {
+	START = 	0,
+	END =   	1,
+	KEYFRAME = 	2
+}
